refactor: extract round types
