+++ src/features/game/index.ts
@@ -3,12 +3,8 @@ export { default as StageControls } from "./ui/controls/StageControls";
 export { default as HintControls } from "./ui/controls/HintControls";
 export { default as LevelsState } from "./model/LevelsState";
 export { default as HintSettings } from "./model/HintSettings";
-export {
-  default as RoundState,
-  type Stage,
-  type Painting,
-} from "./model/RoundState";
-export { StageStatus } from "./model/StageStatus";
+export { default as RoundState } from "./model/round/RoundState";
 export { default as AssembleRow } from "./ui/rows/AssembleRow";
 export { default as PickRow } from "./ui/rows/PickRow";
 export { default as WordCard } from "./ui/card/WordCard";
+export { type Stage, StageStatus } from "./model/round/Stage";
