+++ src/features/game/ui/card/WordCard.ts
@@ -1,7 +1,7 @@
 import { Component, div, span } from "../../../../shared/ui";
 import Draggable from "../../../../entities/drag-n-drop";
 
-import RoundState, { type MoveCardAction } from "../../model/RoundState";
+import RoundState, { type MoveCardAction } from "../../model/round/RoundState";
 import HintSettings from "../../model/HintSettings";
 
 import { type RowData } from "../rows/Row";
