+++ src/features/game/ui/rows/AssembleRow.ts
@@ -1,10 +1,10 @@
 import Row from "./Row";
 import { RowType } from "./RowType";
-import RoundState from "../../model/RoundState";
+import RoundState from "../../model/round/RoundState";
 import HintSettings from "../../model/HintSettings";
 
 import { Publisher } from "../../../../entities/state";
-import { StageStatus } from "../../model/StageStatus";
+import { StageStatus } from "../../model/round/Stage";
 
 import styles from "./Row.module.css";
